[FIX] account_statement_import_online_paypal: Update CONFIGURE, USAGE, and ROADMAP to current Odoo & PayPal menus/procedures/links.
(1) Odoo's menu has changed over the recent versions, and these changes help new users to better understand how to create a PayPal-related Bank Account/Journal and configuration for syncing with PayPal's servers.
(2) The order of CONFIGURE was changed to the logical flow of obtaining PayPal credentials first and then configuring Odoo to match.
(3) PayPal's URL for transaction definitions has changed and was updated in ROADMAP.